eeg-GPT/requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
openai==1.6.1
python-dotenv==1.0.0